Move `Widgets.WidgetBounds` to its own file
